edit user response

+++ shop-backend/src/main/java/com/kma/shop/service/impl/TokenService.java
@@ -29,6 +29,8 @@ public class TokenService {
     }
 
     public void deleteByUser(UserEntity user)   {
+        if(user == null)  return;
+        user.setToken(null);
         tokenRepo.deleteByUser(user);
     }
